PrimitiveAlgoUtils : remove duplication from Curves & Mesh Algos
